Added landing page to App component and moved first question to a separate component
