+++ code/src/components-css/first-question.css
@@ -1,0 +1,4 @@
+.first-question {
+    height: 100vh;
+    background-color: pink;
+}
