+++ code/src/components-css/landing-page.css
@@ -1,0 +1,4 @@
+.landing-page {
+    height: 100vh;
+    background-color: bisque;
+}
