+++ code/src/components/DayPicker.js
@@ -12,25 +12,26 @@ export const DayPicker = (props) => {
   };
 
   return (
-      <>
-    <form onSubmit={(event) => event.preventDefault()}>
-        <div className="day-picker">
-            <label htmlFor="weekdays">Pick a day of the week:</label>
-                <select
-                id="weekdays"
-                onChange={(event) => setDayOption(event.target.value)}
-                value={dayOption} required>
-                    <option value="">Select a day:</option>
-                    <option value="Monday">Monday</option>
-                    <option value="Tuesday">Tuesday</option>
-                    <option value="Wednesday">Wednesday</option>
-                </select>
-            <button onClick={checkInput()}><a href="#final-text">Show me the result!</a></button>
-        </div>
+    <section className="first-question">
+      <form onSubmit={(event) => event.preventDefault()} className="first-question">
+          <div className="day-picker">
+              <label htmlFor="weekdays">Pick a day of the week:</label>
+                  <select
+                  id="weekdays"
+                  onChange={(event) => setDayOption(event.target.value)}
+                  value={dayOption} 
+                  required>
+                      <option value="">Select a day:</option>
+                      <option value="Monday">Monday</option>
+                      <option value="Tuesday">Tuesday</option>
+                      <option value="Wednesday">Wednesday</option>
+                  </select>
+                  <button onClick={checkInput()}><a href="#final-text">Show me the result!</a></button>
+          </div>
 
-        {(visible === true && dayOption) && <FinalText resultColor={props.finalColor} resultNumber={props.chosenNumber} resultDay={dayOption}/>}
-        
-    </form>
-    </>
+          {(visible === true) && <FinalText resultColor={props.finalColor} resultNumber={props.chosenNumber} resultDay={dayOption}/>}
+          
+      </form>
+    </section>
   );
 };
